pyproject.toml
[project]
name = "walmart-refiner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.118.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.34",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
